requirements.txt
flask==2.3.3
flask-cors==4.0.0
reportlab==4.0.4
mysql-connector-python==8.1.0
werkzeug==2.3.7
